# indexForBit optimization