--- a/hamt_bench_test.go
+++ b/hamt_bench_test.go
@@ -3,6 +3,7 @@ package hamt
 import (
 	"context"
 	"encoding/hex"
+	"fmt"
 	"math/rand"
 	"runtime"
 	"testing"
@@ -44,6 +45,40 @@ func BenchmarkSerializeNode(b *testing.B) {
 		if err != nil {
 			b.Fatal(err)
 		}
+	}
+}
+
+type benchSetCase struct {
+	count    int
+	bitwidth int
+}
+
+func BenchmarkSet(b *testing.B) {
+	kCounts := []int{1, 10, 100}
+	bitwidths := []int{5, 8}
+
+	var table []benchSetCase
+	for _, c := range kCounts {
+
+		for _, bw := range bitwidths {
+			table = append(table, benchSetCase{count: c * 1000, bitwidth: bw})
+		}
+
+	}
+	r := rander{rand.New(rand.NewSource(int64(42)))}
+	for _, t := range table {
+		b.Run(fmt.Sprintf("%d/%d", t.count, t.bitwidth), func(b *testing.B) {
+			ctx := context.Background()
+			n := NewNode(NewCborStore(), UseTreeBitWidth(t.bitwidth))
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				for j := 0; j < t.count; j++ {
+					if err := n.Set(ctx, r.randString(), r.randValue()); err != nil {
+						b.Fatal(err)
+					}
+				}
+			}
+		})
 	}
 }