@Stebalien 2020-01-12
Hm. Nevermind, my library assumes that the array length is a multiple of 8.