@jsign 2020-01-17
> Actually, could you add a test case that extracts the current definition of this function into a test case and compares it against the new one? That way, we avoid any regressions.

@Stebalien , I refactored the code to make the test more narrow. Added two tests. 

A randomized one on 32 bytes `BigInt`s which should cover the bitwidth=8. Each case is fully verified in the 256bits that could be asked for. Challenging that test in the `i == len(w)` case (I changed the early return to be +1), it seemed that the randomized one didn't catch up that. Mostly because the randomized fashion + so many bits to cover, make that case a low chance.

Then I created simply a linear one that would be more covering on border cases, but on smaller bit lengths. In this test, that `i == len(w)` case broken-code fails instantly since the linear generation guarantees to cover it.

Kept both since wouldn't hurt.